set up JWT authorization, middleware

+++ routes/events.py
@@ -1,6 +1,7 @@
 from beanie import PydanticObjectId
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, Depends, HTTPException, status
 
+from auth.authenticate import authenticate
 from database.connection import Database
 from models.events import Event, EventUpdate
 from typing import List
@@ -27,14 +28,22 @@ async def retrieve_event(id: int) -> Event:
     return event
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully"
     }
 
 @event_router.put("/update/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
+    
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -44,7 +53,13 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
     return updated_event
 
 @event_router.delete("/delete/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Event not found"
+        )
     event = await event_database.delete(id)
     if event:
         return { "message": "Event deleted successfully" }
